+++ lib/src/async/async_io.rs
@@ -89,6 +89,12 @@ impl AsRef<Chip> for AsyncChip {
     }
 }
 
+impl From<AsyncChip> for Chip {
+    fn from(c: AsyncChip) -> Chip {
+        c.0.into_inner().unwrap()
+    }
+}
+
 /// Async form of [`InfoChangeIterator`].
 ///
 /// Created by [`AsyncChip::info_change_events`].
@@ -266,6 +272,12 @@ impl AsyncRequest {
 impl AsRef<Request> for AsyncRequest {
     fn as_ref(&self) -> &Request {
         self.0.get_ref()
+    }
+}
+
+impl From<AsyncRequest> for Request {
+    fn from(r: AsyncRequest) -> Request {
+        r.0.into_inner().unwrap()
     }
 }